Task #138275 [Sprint 1] Refactor extensions
