+++ src/wsm/wsm/extensions/Dictionary.swift
@@ -1,0 +1,21 @@
+//
+//  Dictionary.swift
+//  wsm
+//
+//  Created by framgia on 9/21/17.
+//  Copyright © 2017 framgia. All rights reserved.
+//
+
+import Foundation
+
+extension Dictionary {
+    mutating func merge(with dictionary: Dictionary) {
+        dictionary.forEach { updateValue($1, forKey: $0) }
+    }
+
+    func merged(with dictionary: Dictionary) -> Dictionary {
+        var dict = self
+        dictionary.forEach { dict.updateValue($1, forKey: $0) }
+        return dict
+    }
+}
