+++ src/wsm/wsm/extensions/Formatter.swift
@@ -1,0 +1,27 @@
+//
+//  NumberFormatter.swift
+//  wsm
+//
+//  Created by framgia on 9/21/17.
+//  Copyright © 2017 framgia. All rights reserved.
+//
+
+import Foundation
+
+struct Formatter {
+    static let decimal = NumberFormatter(numberStyle: .decimal)
+}
+
+extension NumberFormatter {
+    convenience init(numberStyle: NumberFormatter.Style) {
+        self.init()
+        self.numberStyle = numberStyle
+    }
+}
+
+extension DateFormatter {
+    convenience init(format: String) {
+        self.init()
+        self.dateFormat = format
+    }
+}
