+++ src/wsm/wsm/extensions/UITextField.swift
@@ -1,0 +1,13 @@
+//
+//  UITextField.swift
+//  wsm
+//
+//  Created by framgia on 9/21/17.
+//  Copyright © 2017 framgia. All rights reserved.
+//
+
+import Foundation
+
+extension UITextField {
+    var string: String { return text ?? "" }
+}
